Using a custom multi-index map instead.
This fixes the space-leak problem introduced by IxSet.

A blocked indefinitely on STM bug is also fixed. It's caused by a
pattern-matching bug :P

This should be faster... When the load is not too big. In the next step we
might need to start considering implementing some kind of congestion control.